backend/requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
boto==2.39.0
chardet==2.2.1
click==6.7
colorama==0.2.5
DAWG-Python==0.7.2
decorator==3.4.0
docopt==0.6.2
Flask==0.12
Flask-JWT==0.3.2
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.7.2
MarkupSafe==0.18
packaging==16.8
pep8==1.7.0
Pillow==2.3.0
postgres==2.2.1
psycopg2==2.7.1
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-dotenv==0.6.3
pytz==2016.10
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
